Software engineering/ScenarioQualityChecker-Final Project/src/test/java/pl/put/poznan/transformer/logic/scenario/ScenarioElementCounterTest.java
package pl.put.poznan.transformer.logic.scenario;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.isA;
import static org.mockito.Mockito.atLeast;
import static org.mockito.Mockito.atLeastOnce;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoMoreInteractions;
import static org.mockito.Mockito.when;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.junit.jupiter.api.Assertions.assertEquals;

@ExtendWith(MockitoExtension.class)
class ScenarioElementCounterTest {
    @Test
    void testEmpty(@Mock Scenario mockScenario) {
        ScenarioElementCounter counter = new ScenarioElementCounter();
        mockScenario.accept(counter);
        assertEquals(0, counter.getCount());
    }
}